requirements.txt
autopep8
pycodestyle
antlr4-python3-runtime
colored